app/agents/tool.py
from langchain_core.tools import tool
import requests
from bs4 import BeautifulSoup
from tavily import  TavilyClient
import os
from rich import print
import json
from dotenv import load_dotenv
load_dotenv()

HUGGING_FACE_API_KEY = os.getenv("HUGGING_FACE_API_KEY")
tavily =TavilyClient(api_key=os.getenv("TAVILY_API_KEY"))

@tool
def web_search(query : str) -> str:
    """Search the web for recent and reliable information on a topic . Returns Titles , URLs ans snippets."""
    results=tavily.search(query=query,max_results=2)
    
    out = []
    for r in results["results"]:
        out.append(
            f"Title: {r['title']}\n"
            f"URL: {r['url']}\n"
            f"Snippet: {r['content'][:3000]}\n"
        )

    return "\n----\n".join(out)


@tool
def scrape_url(url: str) -> str:
    """Scrape and return clean text content from a given URL for deeper reading."""
    try:
        resp = requests.get(
            url,
            timeout=10,
            headers={"User-Agent": "Mozilla/5.0"}
        )
        resp.raise_for_status()
        soup = BeautifulSoup(resp.text, "html.parser")
        # Remove unwanted tags
        for tag in soup(["script", "style", "nav", "footer", "header"]):
            tag.decompose()
        text = soup.get_text(separator=" ", strip=True)
        # Clean extra spaces
        clean_text = " ".join(text.split())
        return clean_text[:800]  # increased limit
    except Exception as e:
        return f"Could not scrape URL: {str(e)}"
    
