requirements.txt
# Core LangChain ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
langchain-huggingface>=0.0.1

# Hugging Face
huggingface_hub>=0.23.0
transformers>=4.41.0
accelerate>=0.30.0

# Search & Scraping
tavily-python>=0.3.3
beautifulsoup4>=4.12.3
requests>=2.32.0
lxml>=5.2.1

# Async & Networking
httpx>=0.27.0
aiohttp>=3.9.5
uvicorn>=0.27.0
fastapi>=0.110.0

# Data & Utilities
pydantic>=2.7.0
pandas>=2.2.2
orjson>=3.10.3
loguru>=0.7.2
python-dotenv>=1.0.1
python-multipart>=0.0.9

# Database & Auth
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.9
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0

# Vector & Tokens
tiktoken>=0.7.0
faiss-cpu>=1.8.0